// mobile/redux/client/client.types.ts
export enum ClientTypes {
    FETCH_DEVICS_REQUEST = 'FETCH_DEVICS_REQUEST',
    FETCH_DEVICS_SUCCESS = 'FETCH_DEVICS_SUCCESS',
    FETCH_DEVICS_FAILURE = 'FETCH_DEVICS_FAILURE',
    FETCH_USERS_REQUEST = 'FETCH_USERS_REQUEST',
    FETCH_USERS_SUCCESS = 'FETCH_USERS_SUCCESS',
    FETCH_USERS_FAILURE = 'FETCH_USERS_FAILURE',
    UPDATE_DEVICE = 'UPDATE_DEVICE',
    FILTER_DEVICE_BY_NAME = 'FILTER_DEVICE_BY_NAME',
}